Move suffixed_number_value to kubernetes_tools, delete the rest of metastatus_lib.py
